requirements.txt
Click==7.0
paho-mqtt==1.4.0
